# doc/03_openapi_koipa_kl.yaml
openapi: 3.0.3
info:
  title: Koipa AI Engine API
  description: |
    한국지식재산보호원 AI 영업비밀관리시스템 — 한국지식재산보호원 AI 엔진의
    KL 웹시스템 연동용 REST API 명세.

    담당 기능: FUN-003, FUN-004, FUN-005, FUN-022, FUN-023, FUN-024
    호출 주체: KL 웹시스템 (Backend)
    제공 주체: Koipa AI Engine (FastAPI)

    인증: API Key (서버 간 통신) + 선택적 JWT (관리자 행위 추적)
    모든 응답은 application/json. 파일 업로드는 multipart/form-data.
    날짜·시각은 ISO 8601 UTC. 모든 ID는 UUIDv4.

    내부 RAG 백엔드: PostgreSQL pgvector dense + bigram tsvector `ts_rank`
    하이브리드. 본 명세는 backend-agnostic 하므로 KL 측은 RAG 동작 세부를
    몰라도 호출 가능. ES 경로는 레거시 어댑터로만 남아 있다.
  version: 1.0.0-draft
  contact:
    name: Koipa AI Engineering
servers:
  - url: http://koipa-api:8000/api/v1
    description: 온프레미스 Docker 내부망
  - url: https://{host}/api/v1
    description: 프로덕션
    variables:
      host:
        default: ai.example.koipa.local
security:
  - ApiKeyAuth: []
...
paths:
...
  /admin/model/reload:
    post:
      tags: [admin]
      summary: 활성 모델 핫리로드
      responses:
        '200':
          description: 리로드 결과
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
  /admin/model/activate:
    post:
      tags: [admin]
      summary: 특정 ModelVersion 수동 활성화
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version_label]
              properties:
                version_label: { type: string }
                force: { type: boolean, default: false }
      responses:
        '200':
          description: 활성화 또는 게이트 차단 결과
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
  /admin/escalation-held:
    get:
      tags: [admin]
      summary: 이중검토 보류 건수
      responses:
        '200':
          description: 등급별 보류 건수
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
  /admin/locked-readiness:
    get:
      tags: [admin]
      summary: locked_gold_eval readiness
      responses:
        '200':
          description: 등급별 locked 평가 준비도
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
  /admin/dashboard:
    get:
      tags: [admin]
      summary: 운영 검수 대시보드 종합
      responses:
        '200':
          description: 운영 상태 종합
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
...
  /admin/audit-log:
    get:
      tags: [admin]
      summary: 감사 로그 조회 — 누가 언제 무엇을 했는가
      description: >-
        헬스체크·정적자산은 기록에서 제외된다. 응답의 actorless_ratio 는 행위자(actor_id)가
        없는 기록 비율 — 공유 API 키 모드에서는 개별 신원이 남지 않는다.
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 50 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
        - { in: query, name: action, schema: { type: string }, description: 부분일치 }
        - { in: query, name: success, schema: { type: boolean } }
        - { in: query, name: actor_id, schema: { type: string } }
      responses:
        '200':
          description: 감사 로그 페이지
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
  /admin/model/rollback:
    post:
      tags: [admin]
      summary: 활성 모델을 직전 활성본으로 롤백 + 무중단 리로드
      description: >-
        배포 게이트를 적용하지 않는다 — 회귀를 되돌리는 방향이라 막으면 사고 시 복구가
        불가능해진다. 대신 사유(reason)가 필수이며 활성화와 동일한 advisory lock 으로
        직렬화된다. 복귀 후보가 없으면 rolled_back=false 로 응답한다(리로드 안 함).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, description: 감사에 남는 유일한 맥락이므로 필수 }
      responses:
        '200':
          description: 롤백 결과 (rolled_back · reason)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '403': { description: "admin role required" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...